Retrieve all the youtube playlist
The IDs list is retrieved via youtube-dl.
Then the songs information is retrieved via oembed youtube link. A json
format result is returned. The title is in Unicode format and needs to be
converted to UTF-8.
Now that the list is updated later after the first display, we must check
whether the graphic objects exist.